refactor: Move `CompassDirection` into a top-level enum
Preparatory work to make it easier to share between different implementations.
